% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_word_list.R
\name{access_word_list}
\alias{access_word_list}
\title{Retrieve internal word list.}
\usage{
access_word_list(word_type)
}
\arguments{
\item{word_type}{Name of word list to retrieve (Character)}
}
\value{
word list (Character Vector).
}
\description{
Access an internal word list. See \code{Word types} below.
}
\section{Word types}{


Word lists have been collected from the internet. They might contain
words that seem out of place. Let me know and I'll remove them.

\subsection{Positive words}{
[superb]
[positive_adj]
[positive_vb]
[positive_vb_past]
[positive_adv]}

\subsection{Verb forms}{
[verb_base]
[verb_past]
[verb_past_participle]
[verb_es]
[verb_ing]}

\subsection{Noun forms}{
[noun]
[noun_plural]}

\subsection{Persons}{
[salutation]
[male_first_name]
[female_first_name]
[male_full_name]
[female_full_name]}

\subsection{Specific synonyms}{
Notice: These are added as needed by author.

[create_syn]
[small_syn]
[big_syn]}

\subsection{Nationality}{
[nationality_adj]
[nationality_people]}

\subsection{Statistics}{
[statistics_tradition]}
}

\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
