% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{bhappyr}
\alias{bhappyr}
\alias{bhappyr-package}
\title{bhappyr: A package for generating amusing sentences.}
\description{
I will add a description later...
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
