% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{big_syn}
\alias{big_syn}
\title{Synonyms for "big"}
\format{A vector with 16 elements}
\usage{
big_syn
}
\description{
A character vector containing English synonyms for "big".
}
\keyword{datasets}
