% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{create_syn}
\alias{create_syn}
\title{Synonyms for "create"}
\format{A vector with 22 elements}
\usage{
create_syn
}
\description{
A character vector containing English synonyms for "create".
}
\keyword{datasets}

