% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{female_first_name}
\alias{female_first_name}
\title{Female first names}
\format{A vector with 159 elements}
\usage{
female_first_name
}
\description{
A character vector containing female first names.
}
\keyword{datasets}
