% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{female_full_name}
\alias{female_full_name}
\title{Female full names}
\format{A vector with 150 elements}
\usage{
female_full_name
}
\description{
A character vector containing female full names.
}
\keyword{datasets}
