% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHappyFeedback.R
\name{generateHappyFeedback}
\alias{generateHappyFeedback}
\title{Generate happy feedback.}
\usage{
generateHappyFeedback()
}
\value{
Cat() statement
}
\description{
Addin for RStudio for generating amusing feedback to lighten the mood
of statisticians.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to:

\code{Tools} >> \code{Addins} >> \code{Browse Addins} >> \code{Keyboard Shortcuts}.

Find \code{"Become Happy"} and press its field under \code{Shortcut}.

Press desired key command.

Press \code{Apply}.

Press \code{Execute}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
