% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{male_first_name}
\alias{male_first_name}
\title{Male first names}
\format{A vector with 159 elements}
\usage{
male_first_name
}
\description{
A character vector containing male first names.
}
\keyword{datasets}
