% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{male_full_name}
\alias{male_full_name}
\title{Male full names}
\format{A vector with 150 elements}
\usage{
male_full_name
}
\description{
A character vector containing male full names.
}
\keyword{datasets}
