% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{positive_adj}
\alias{positive_adj}
\title{Positive adjectives}
\format{A vector with 124 elements}
\usage{
positive_adj
}
\description{
A character vector containing positive English adjectives.
}
\keyword{datasets}
