% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{positive_adv}
\alias{positive_adv}
\title{Positive adverbs}
\format{A vector with 861 elements}
\usage{
positive_adv
}
\description{
A character vector containing positive English adverbs.
}
\keyword{datasets}
