% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{positive_vb}
\alias{positive_vb}
\title{Positive verbs}
\format{A vector with 916 elements}
\usage{
positive_vb
}
\description{
A character vector containing positive English verbs.
}
\keyword{datasets}
