% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{positive_vb_past}
\alias{positive_vb_past}
\title{Positive verbs with past tense}
\format{A vector with 500 elements}
\usage{
positive_vb_past
}
\description{
A character vector containing positive English past tense verbs.
}
\keyword{datasets}
