% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_sentence.R
\name{rate_sentence}
\alias{rate_sentence}
\title{Generate and rate sentence.}
\usage{
rate_sentence(n = 1, question = "Was this sentence funny?", rscale = c(1,
  10), max_rating = "Extremely", min_rating = "Not at all!",
  change_label = FALSE, change_rscale = FALSE, rscale_limits = c(1, 1000),
  simulate = FALSE, structures = NULL, word_lists = NULL,
  int_structures = TRUE, int_word_lists = TRUE)
}
\arguments{
\item{n}{Number of sentences to generate (Integer).}

\item{question}{(Character)}

\item{rscale}{Rating scale min and max.

E.g. \code{c(1,10)}

Pass \code{'random'} to automatically create
scale}

\item{max_rating}{Label for max rating value}

\item{min_rating}{Label for min rating value}

\item{change_label}{Whether to sample new word for labels,
which have been passed as word types, on each run
or only once (Logical).}

\item{change_rscale}{Whether to create new random scale
on each run or only once (Logical).}

\item{rscale_limits}{Limits if generating random scale

E.g. c(1,1000)}

\item{simulate}{Simulate user ratings (Logical).}

\item{structures}{Character Vector. E.g.:

\code{c("My [noun] is [positive_adj] today!",
"[salutation]! How you've grown with the task!")}

Constrain the selection of words by specifying \strong{word info codes}.
E.g. only sample elements from the word list with
max. 1 word in them, i.e. no white spaces:

\code{"[salutation(w1)]"}

Pass multiple word types - from which one is randomly selected:

\code{"[male_first_name/female_first_name]"}

See available \code{Word Types} below.

See available \code{Word Info Codes} below.}

\item{word_lists}{List of word lists. A word list
is named by the word type, e.g. verb.}

\item{int_structures}{Use internal structures (Logical).}

\item{int_word_lists}{Use internal word lists (Logical).}
}
\value{
Collected responses, sentences, rating labels and
and the rating scale (Dataframe).
}
\description{
Generates one or more sentences and asks user
to rate them in the console.

Pass question, rating scale, and min/max rating labels.

Randomly create rating scales - either once or for each sentence.

Min/max rating labels can be passed as word types. Words are sampled
either once or once per sentence.
See \code{Word types} below.
}
\section{Word types}{


Word lists have been collected from the internet. They might contain
words that seem out of place. Let me know and I'll remove them.

\subsection{Positive words}{
[superb]
[positive_adj]
[positive_vb]
[positive_vb_past]
[positive_adv]}

\subsection{Verb forms}{
[verb_base]
[verb_past]
[verb_past_participle]
[verb_es]
[verb_ing]}

\subsection{Noun forms}{
[noun]
[noun_plural]}

\subsection{Persons}{
[salutation]
[male_first_name]
[female_first_name]
[male_full_name]
[female_full_name]}

\subsection{Specific synonyms}{
Notice: These are added as needed by author.

[create_syn]
[small_syn]
[big_syn]}

\subsection{Nationality}{
[nationality_adj]
[nationality_people]}

\subsection{Statistics}{
[statistics_tradition]}
}

\examples{
# Attach package
library(bhappyr)

# Notice: The argument 'simulate' is set to TRUE
# as the function is elsewise interactive, i.e.
# it is waiting for user input.
# Remove or set to FALSE when running as experiment!

# Example 1
# Generate 1 sentence and rate it.
rate_sentence(n=1, simulate = TRUE)

# Example 2
# Generate 3 sentences and rate them.
# Use random scale on each run.
rate_sentence(n=3, rscale = 'random',
              change_rscale = TRUE,
              simulate = TRUE)

# Example 3
# Generate 3 sentences and rate them.
# Use random max_rating label on each run.
rate_sentence(n=3, max_rating = "[superb(upper)]",
              change_label = TRUE,
              simulate = TRUE)
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
