% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{small_syn}
\alias{small_syn}
\title{Synonyms for "small"}
\format{A vector with 19 elements}
\usage{
small_syn
}
\description{
A character vector containing English synonyms for "small".
}
\keyword{datasets}

